move signing logic to remote functions
