+++ src/routes/(app)/expenses/[id]/+page.server.ts
@@ -1,22 +1,16 @@
-import apiNames from "$lib/utils/apiNames";
-import { isAuthorized } from "$lib/utils/authorization";
 import { redirect } from "sveltekit-flash-message/server";
 import { error } from "@sveltejs/kit";
 import { fail, message, superValidate } from "sveltekit-superforms";
 import { zod4 } from "sveltekit-superforms/adapters";
-import { z } from "zod";
 import { getCostCenter } from "../config";
 import { expensesInclusion } from "../getExpenses";
 import { sendNotificationToSigner } from "../helper";
-import { sendExpenseToBookkeeping } from "$lib/expenses/sendToBookkeeping";
 import {
   getSigner,
   resolveSignerLogic,
   updateSignersCacheIfNecessary,
 } from "../signers";
 import { updateExpenseSchema, updateItemSchema } from "../types";
-import { setFlash } from "sveltekit-flash-message/server";
-import { convertPriceToCents } from "$lib/utils/convertPrice";
 
 export const load = async ({ locals, params }) => {
   const { prisma } = locals;
@@ -38,10 +32,6 @@ export const load = async ({ locals, params }) => {
     updateForm: await superValidate(expense, zod4(updateExpenseSchema)),
   };
 };
-
-const approveReceiptSchema = z.object({
-  itemId: z.string(),
-});
 
 export const actions = {
   updateReceipt: async ({ params, request, locals }) => {
@@ -147,161 +137,5 @@ export const actions = {
       },
       event,
     );
-  },
-  disapproveReceipt: async (event) => {
-    const { locals, request } = event;
-    const { prisma, user } = locals;
-    if (!user?.memberId)
-      throw fail(401, {
-        fail,
-        message: "Du måste vara inloggad för att av-godkänna utlägg",
-      });
-    const form = await superValidate(request, zod4(approveReceiptSchema));
-    if (!form.valid) return fail(400, { form });
-    const canAlwaysSign = isAuthorized(apiNames.EXPENSES.CERTIFICATION, user);
-    try {
-      await prisma.expenseItem.update({
-        where: {
-          id: form.data.itemId,
-          OR: canAlwaysSign
-            ? undefined
-            : [
-                {
-                  signedByMemberId: user.memberId,
-                },
-                {
-                  signerMemberId: user.memberId,
-                },
-              ],
-          signedAt: {
-            not: null,
-          },
-        },
-        data: {
-          signedByMemberId: null,
-          signedAt: null,
-        },
-      });
-    } catch {
-      return message(form, {
-        message: "Kunde inte av-godkänna kvitto",
-        type: "error",
-      });
-    }
-    return message(form, {
-      message: "Kvitto av-godkänt",
-      type: "success",
-    });
-  },
-  approveReceipt: async (event) => {
-    const { locals, params, request } = event;
-    const { prisma, user } = locals;
-    if (!user?.memberId)
-      throw error(401, "Du måste vara inloggad för att godkänna utlägg");
-    if (params.id.length === 0 || Number.isNaN(Number(params.id)))
-      throw error(404, "Expense id should be a number");
-    const form = await superValidate(request, zod4(approveReceiptSchema));
-    if (!form.valid) return fail(400, { form });
-    const canAlwaysSign = isAuthorized(apiNames.EXPENSES.CERTIFICATION, user);
-    try {
-      await prisma.expenseItem.update({
-        where: {
-          id: form.data.itemId,
-          expenseId: Number(params.id),
-          signedByMemberId: null,
-          signedAt: null,
-          signerMemberId: canAlwaysSign ? undefined : user.memberId, // only sign items which are assigned to the user, or all if user is allowed to sign all (i.e. is the treasurer or president)
-        },
-        data: {
-          signedByMemberId: user.memberId,
-          signedAt: new Date(),
-        },
-      });
-    } catch {
-      return message(form, {
-        message: "Kunde inte godkänna kvitto",
-        type: "error",
-      });
-    }
-    return message(form, {
-      message: "Kvitto godkänd",
-      type: "success",
-    });
-  },
-  approveAll: async (event) => {
-    const { locals, params } = event;
-    const { prisma, user } = locals;
-    if (!user?.memberId)
-      throw error(401, "Du måste vara inloggad för att godkänna utlägg");
-    if (params.id.length === 0 || Number.isNaN(Number(params.id)))
-      throw error(404, "Expense id should be a number");
-    const canAlwaysSign = isAuthorized(apiNames.EXPENSES.CERTIFICATION, user);
-    const result = await prisma.expenseItem.updateMany({
-      where: {
-        expenseId: Number(params.id),
-        signedByMemberId: null,
-        signedAt: null,
-        signerMemberId: canAlwaysSign ? undefined : user.memberId, // only sign items which are assigned to the user, or all if user is allowed to sign all (i.e. is the treasurer or president)
-      },
-      data: {
-        signedByMemberId: user.memberId,
-        signedAt: new Date(),
-      },
-    });
-    if (result.count === 0) {
-      throw error(400, "Utlägget kunde inte godkännas");
-    }
-    return redirect(
-      "/expenses",
-      {
-        message: "Utlägg godkänt",
-        type: "success",
-      },
-      event,
-    );
-  },
-  sendToBookkeeping: async (event) => {
-    const { locals, params } = event;
-    const { prisma, user } = locals;
-
-    if (!user?.memberId) {
-      throw error(
-        401,
-        "Du måste vara inloggad för att skicka utlägg till bokföring",
-      );
-    }
-
-    if (!isAuthorized(apiNames.EXPENSES.BOOKKEEPING, user)) {
-      throw error(
-        403,
-        "Du har inte behörighet att skicka utlägg till bokföring",
-      );
-    }
-
-    if (params.id.length === 0 || Number.isNaN(Number(params.id))) {
-      throw error(404, "Expense id should be a number");
-    }
-
-    try {
-      await sendExpenseToBookkeeping(prisma, Number(params.id));
-
-      return redirect(
-        "/expenses/all",
-        {
-          message: "Utlägg skickat till bokföring",
-          type: "success",
-        },
-        event,
-      );
-    } catch (e) {
-      setFlash(
-        {
-          message: e instanceof Error ? e.message : "Ett fel uppstod",
-          type: "error",
-        },
-        event,
-      );
-      return fail(400);
-    }
   },
 };
